Rework GNOME Split constructor.
Use few methods to build the main object by step.
In this way, it will be easier to tweak the code
and to search for errors.

+++ src/org/gnome/split/gtk/MainWindow.java
@@ -330,7 +330,7 @@ public class MainWindow extends Window implements Window.DeleteEvent
      */
     public void selectDefaultView() {
         // Get the default view
-        int view = app.getConfig().DEFAULT_VIEW;
+        byte view = app.getConfig().DEFAULT_VIEW;
 
         // The merge view is the default one
         if (view == 1) {
